feat: add single cmd processing
