+++ Makefile
@@ -6,7 +6,7 @@
 #    By: sqiu <sqiu@student.42vienna.com>           +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2023/07/28 13:03:05 by gwolf             #+#    #+#              #
-#    Updated: 2023/08/07 16:57:36 by sqiu             ###   ########.fr        #
+#    Updated: 2023/08/08 12:33:06 by sqiu             ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -107,6 +107,7 @@ SRC :=	utils_memory.c \
 		executor_child.c \
 		executor_cleanup.c \
 		executor_utils.c \
+		executor_utils2.c \
 		executor_heredoc.c \
 		executor_builtin.c \
 		cleanup.c
