+++ inc/mod_executor.h
@@ -6,7 +6,7 @@
 /*   By: sqiu <sqiu@student.42vienna.com>           +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/07/31 11:04:32 by sqiu              #+#    #+#             */
-/*   Updated: 2023/08/08 11:16:05 by sqiu             ###   ########.fr       */
+/*   Updated: 2023/08/08 12:32:29 by sqiu             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,9 +22,8 @@
 /* ====== Functions ====== */
 
 // cmd handling
-void	ft_init_exec(t_cmd *cmd);
-t_err	ft_create_pipes(t_cmd *cmd);
-t_err	ft_execute_cmds(t_cmd *cmd, char **envp, char **paths, t_data *data);
+t_err	ft_execute_scmd(t_cmd *cmd, char **envp, char **paths, t_data *data);
+t_err	ft_execute_pcmds(t_cmd *cmd, char **envp, char **paths, t_data *data);
 t_err	ft_process_cmd(t_cmd *cmd, t_err err, char **envp, t_data *data);
 
 // create child processes
@@ -52,6 +51,8 @@ t_err	ft_close(int fd);
 t_err	ft_plug_pipe(t_cmd *cmd, bool close_read_end);
 
 // utils
+void	ft_init_exec(t_cmd *cmd);
+t_err	ft_create_pipes(t_cmd *cmd);
 t_err	ft_check_cmd_access(char **args, char **cmd_paths);
 t_err	ft_prefix_path(char **args, char **cmd_paths);
 t_err	ft_get_path(char **envp, char ***paths);
